seo-agent/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
anthropic>=0.40.0
